projects/whale-tracker/requirements.txt
discord.py==2.3.2
aiohttp==3.9.1
requests==2.31.0
python-dotenv==1.0.0
supabase==2.7.1
web3==6.11.1
stripe==10.0.0
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.5.0
